+++ src/main/kotlin/com/github/mohamead/spiderlog/plugin/action/OpenFileAction.kt
@@ -1,21 +1,21 @@
 package com.github.mohamead.spiderlog.plugin.action
 
+import com.github.mohamead.spiderlog.plugin.ui.SpiderlogPanel
 import com.github.mohamead.spiderlog.plugin.util.LogTracer
 import com.github.mohamead.spiderlog.plugin.util.clearContent
-import com.github.mohamead.spiderlog.plugin.util.getToolWindowPanel
 import com.github.mohamead.spiderlog.plugin.util.openPath
 import com.github.mohamead.spiderlog.plugin.util.validExtension
 import com.intellij.openapi.actionSystem.AnActionEvent
 import java.awt.EventQueue
+import javax.swing.Icon
 
-internal class OpenFileAction : SpiderlogAction() {
+internal class OpenFileAction(val spiderlogPanel: SpiderlogPanel, override val icon: Icon?) : SpiderlogAction(icon) {
 
     override fun actionPerformed(e: AnActionEvent) {
         val extension = validExtension.joinToString("|")
         val filePath = openPath(e.project!!, "Open file", "Select a file (${extension})") ?: return
-        val toolWindowPanel = getToolWindowPanel(e.project!!)
-        toolWindowPanel.table.clearContent()
-        EventQueue.invokeLater { LogTracer().display(toolWindowPanel, filePath.toFile()) }
+        spiderlogPanel.table.clearContent()
+        EventQueue.invokeLater { LogTracer().display(spiderlogPanel, filePath.toFile()) }
     }
 
 }