+++ src/main/kotlin/com/github/mohamead/spiderlog/plugin/enum/FontStyle.kt
@@ -8,7 +8,7 @@ internal enum class FontStyle(val index: Int) {
     companion object {
 
         internal fun findByIndex(index: Int): FontStyle {
-            return values().firstOrNull { it.index == index }
+            return entries.firstOrNull { it.index == index }
                 ?: throw NoSuchElementException("FontStyle with index '$index' not found.")
         }