+++ src/main/kotlin/com/github/mohamead/spiderlog/plugin/icons/SpiderlogIcons.kt
@@ -5,6 +5,6 @@ import com.intellij.openapi.util.IconLoader
 internal object SpiderlogIcons {
 
     @JvmField
-    val spiderToolWindowIcon = IconLoader.getIcon("/icons/toolWindowIcon.svg", javaClass)
+    val spiderlog = IconLoader.getIcon("/icons/spiderlog.svg", javaClass)
 
 }