+++ src/main/kotlin/com/github/mohamead/spiderlog/plugin/ui/BaseSpiderlogPanel.kt
@@ -1,61 +1,42 @@
 package com.github.mohamead.spiderlog.plugin.ui
 
 import com.github.mohamead.spiderlog.plugin.setting.SettingState
-import com.github.mohamead.spiderlog.plugin.util.getActionGroup
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.ActionManager
 import com.intellij.openapi.actionSystem.ActionToolbar
+import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.openapi.ide.CopyPasteManager
-import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.ui.JBMenuItem
 import com.intellij.openapi.ui.JBPopupMenu
 import com.intellij.openapi.ui.SimpleToolWindowPanel
-import com.intellij.ui.components.JBScrollPane
 import com.intellij.ui.table.JBTable
 import java.awt.BorderLayout
 import java.awt.Font
-import java.awt.Robot
 import java.awt.datatransfer.StringSelection
-import java.awt.event.KeyEvent
-import javax.swing.JProgressBar
 
-internal class SpiderlogToolWindowPanel : SimpleToolWindowPanel(false), DumbAware {
+internal abstract class BaseSpiderlogPanel : SimpleToolWindowPanel(false) {
 
-    internal val table: JBTable
-    internal val progressBar: JProgressBar
-
-    init {
-        this.toolbar = buildToolbar().component
-        this.table = buildTable()
-        this.progressBar = JProgressBar()
-        val panel = buildPanel().apply {
-            add(progressBar)
-            add(JBScrollPane(table))
-        }
-        this.add(panel)
-    }
-
-    private fun buildToolbar(): ActionToolbar {
+    protected fun buildToolbar(defaultActionGroup: DefaultActionGroup): ActionToolbar {
         return ActionManager.getInstance()
-            .createActionToolbar("Spiderlog Navigator Toolbar", getActionGroup("Spiderlog.NavigatorToolbar"), false)
+            .createActionToolbar("Spiderlog Navigator Toolbar", defaultActionGroup, false)
             .apply {
                 layoutPolicy = ActionToolbar.AUTO_LAYOUT_POLICY
                 adjustTheSameSize(true)
                 setShowSeparatorTitles(true)
-                targetComponent = this@SpiderlogToolWindowPanel
+                targetComponent = this@BaseSpiderlogPanel
             }
     }
 
-    private fun buildPanel(): SimpleToolWindowPanel {
+    protected fun buildPanel(): SimpleToolWindowPanel {
         return SimpleToolWindowPanel(true, true).apply {
             layout = BorderLayout()
         }
     }
 
-    private fun buildTable(): JBTable {
+    protected fun buildTable(): JBTable {
         return JBTable().apply {
             font = buildFont()
-            componentPopupMenu = buildPopupMenu()
+            componentPopupMenu = buildPopupMenu(this)
             tableHeader.reorderingAllowed = false
             tableHeader.resizingAllowed = false
             inheritsPopupMenu = true
@@ -64,29 +45,28 @@ internal class SpiderlogToolWindowPanel : SimpleToolWindowPanel(false), DumbAwar
         }
     }
 
-    private fun buildFont(): Font {
+    protected fun buildFont(): Font {
         val state = SettingState().getInstance().state
         return Font(state.fontName, state.fontStyle.index, state.fontSize)
     }
 
-    private fun buildPopupMenu(): JBPopupMenu {
+    private fun buildPopupMenu(table: JBTable): JBPopupMenu {
         return JBPopupMenu().apply {
-            val popupMenuCopyMenuItem = JBMenuItem("Copy", AllIcons.Actions.Copy).apply {
+            val copyMenuItem = JBMenuItem("Copy", AllIcons.Actions.Copy).apply {
                 this.addActionListener {
-                    val selectedRow = table.selectedRows
+                    val selectedRows = table.selectedRows
                     val selectedColumns = table.selectedColumns
-                    if (selectedRow.isNotEmpty() && selectedColumns.isNotEmpty()) {
-
+                    if (selectedRows.isNotEmpty() && selectedColumns.isNotEmpty()) {
+                        var textToCopy = ""
+                        for (row in selectedRows) {
+                            textToCopy += table.getValueAt(row, 0).toString() + "\n"
+                        }
+                        CopyPasteManager.getInstance().setContents(StringSelection(textToCopy.trim().trimIndent()))
                     }
                 }
             }
-            this.add(popupMenuCopyMenuItem)
+            this.add(copyMenuItem)
         }
-    }
-
-    internal fun updateUi() {
-        table.font = buildFont()
-        table.updateUI()
     }
 
 }