+++ src/main/kotlin/com/github/mohamead/spiderlog/plugin/util/LogTracer.kt
@@ -1,6 +1,6 @@
 package com.github.mohamead.spiderlog.plugin.util
 
-import com.github.mohamead.spiderlog.plugin.ui.SpiderlogToolWindowPanel
+import com.github.mohamead.spiderlog.plugin.ui.SpiderlogPanel
 import java.beans.PropertyChangeEvent
 import java.io.File
 import javax.swing.SwingWorker
@@ -8,13 +8,13 @@ import javax.swing.table.DefaultTableModel
 
 internal class LogTracer {
 
-    internal fun display(toolWindowPanel: SpiderlogToolWindowPanel, file: File) {
+    internal fun display(spiderlogPanel: SpiderlogPanel, file: File) {
         val model = DefaultTableModel()
-        toolWindowPanel.table.model = model
+        spiderlogPanel.table.model = model
         val worker = LogWorker(file, model)
         worker.addPropertyChangeListener { e: PropertyChangeEvent ->
             val state: SwingWorker.StateValue = e.newValue as SwingWorker.StateValue
-            toolWindowPanel.progressBar.isIndeterminate = state == SwingWorker.StateValue.STARTED
+            spiderlogPanel.progressBar.isIndeterminate = state == SwingWorker.StateValue.STARTED
         }
         worker.execute()
     }